[Network Path] macOS support
