+++ cmd/system-probe/modules/traceroute_others.go
@@ -1,0 +1,20 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// go:build windows || darwin
+
+package modules
+
+import (
+	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
+	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
+)
+
+// Traceroute is a factory for NDMs Traceroute module
+var Traceroute = module.Factory{
+	Name:             config.TracerouteModule,
+	ConfigNamespaces: tracerouteConfigNamespaces,
+	Fn:               createTracerouteModule,
+}
