+++ pkg/process/net/common.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux || windows
+//go:build linux || windows || darwin
 
 package net
 
@@ -18,15 +18,10 @@ import (
 	"sync"
 	"time"
 
-	model "github.com/DataDog/agent-payload/v5/process"
 	"google.golang.org/protobuf/proto"
 
 	"github.com/DataDog/datadog-agent/pkg/languagedetection/languagemodels"
-	netEncoding "github.com/DataDog/datadog-agent/pkg/network/encoding/unmarshal"
-	procEncoding "github.com/DataDog/datadog-agent/pkg/process/encoding"
-	reqEncoding "github.com/DataDog/datadog-agent/pkg/process/encoding/request"
 	languagepb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/languagedetection"
-	pbgo "github.com/DataDog/datadog-agent/pkg/proto/pbgo/process"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/retry"
 )
@@ -87,81 +82,6 @@ func GetRemoteSystemProbeUtil(path string) (*RemoteSysProbeUtil, error) {
 	}
 
 	return globalUtil, nil
-}
-
-// GetProcStats returns a set of process stats by querying system-probe
-func (r *RemoteSysProbeUtil) GetProcStats(pids []int32) (*model.ProcStatsWithPermByPID, error) {
-	procReq := &pbgo.ProcessStatRequest{
-		Pids: pids,
-	}
-
-	reqBody, err := reqEncoding.GetMarshaler(reqEncoding.ContentTypeProtobuf).Marshal(procReq)
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequest("POST", procStatsURL, bytes.NewReader(reqBody))
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Accept", contentTypeProtobuf)
-	req.Header.Set("Content-Type", procEncoding.ContentTypeProtobuf)
-	resp, err := r.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("proc_stats request failed: Probe Path %s, url: %s, status code: %d", r.path, procStatsURL, resp.StatusCode)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	contentType := resp.Header.Get("Content-type")
-	results, err := procEncoding.GetUnmarshaler(contentType).Unmarshal(body)
-	if err != nil {
-		return nil, err
-	}
-
-	return results, nil
-}
-
-// GetConnections returns a set of active network connections, retrieved from the system probe service
-func (r *RemoteSysProbeUtil) GetConnections(clientID string) (*model.Connections, error) {
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s?client_id=%s", connectionsURL, clientID), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Accept", contentTypeProtobuf)
-	resp, err := r.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("conn request failed: Probe Path %s, url: %s, status code: %d", r.path, connectionsURL, resp.StatusCode)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	contentType := resp.Header.Get("Content-type")
-	conns, err := netEncoding.GetUnmarshaler(contentType).Unmarshal(body)
-	if err != nil {
-		return nil, err
-	}
-
-	return conns, nil
 }
 
 // GetPing returns the results of a ping to a host
