+++ pkg/process/net/common_linux_windows.go
@@ -1,0 +1,83 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux || windows
+
+package net
+
+// GetProcStats returns a set of process stats by querying system-probe
+func (r *RemoteSysProbeUtil) GetProcStats(pids []int32) (*model.ProcStatsWithPermByPID, error) {
+	procReq := &pbgo.ProcessStatRequest{
+		Pids: pids,
+	}
+
+	reqBody, err := reqEncoding.GetMarshaler(reqEncoding.ContentTypeProtobuf).Marshal(procReq)
+	if err != nil {
+		return nil, err
+	}
+
+	req, err := http.NewRequest("POST", procStatsURL, bytes.NewReader(reqBody))
+	if err != nil {
+		return nil, err
+	}
+
+	req.Header.Set("Accept", contentTypeProtobuf)
+	req.Header.Set("Content-Type", procEncoding.ContentTypeProtobuf)
+	resp, err := r.httpClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("proc_stats request failed: Probe Path %s, url: %s, status code: %d", r.path, procStatsURL, resp.StatusCode)
+	}
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	contentType := resp.Header.Get("Content-type")
+	results, err := procEncoding.GetUnmarshaler(contentType).Unmarshal(body)
+	if err != nil {
+		return nil, err
+	}
+
+	return results, nil
+}
+
+// GetConnections returns a set of active network connections, retrieved from the system probe service
+func (r *RemoteSysProbeUtil) GetConnections(clientID string) (*model.Connections, error) {
+	req, err := http.NewRequest("GET", fmt.Sprintf("%s?client_id=%s", connectionsURL, clientID), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	req.Header.Set("Accept", contentTypeProtobuf)
+	resp, err := r.httpClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("conn request failed: Probe Path %s, url: %s, status code: %d", r.path, connectionsURL, resp.StatusCode)
+	}
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	contentType := resp.Header.Get("Content-type")
+	conns, err := netEncoding.GetUnmarshaler(contentType).Unmarshal(body)
+	if err != nil {
+		return nil, err
+	}
+
+	return conns, nil
+}
